added all api documentation under swagger file and swagger apis for testing

+++ app/src/controllers/cartController.ts
@@ -20,7 +20,7 @@ export let saveCart = async (req: Request, res: Response) => {
     baseResponse.response = result
   } catch (e) {
     baseResponse.isSuccess = false
-    baseResponse.response = JSON.stringify(e)
+    baseResponse.response = e
   }
   res.send(baseResponse)
 }
@@ -36,16 +36,27 @@ export let addItem = async (req: Request, res: Response) => {
 
   try {
     cartEntity = await cartRepo.getCartById(req.body.cart_id)
-    cartRepo.addItem(cartEntity, req.body.item)
-    cartEntity = cartRepo.updateCart(cartEntity)
-    let result = await cartRepo.saveCart(cartEntity)
-    baseResponse.isSuccess = true
-    baseResponse.response = result
+    if (cartEntity != undefined) {
+      if (req.body.item == undefined) {
+        baseResponse.responseCode = 400
+        baseResponse.message = 'invalid item'
+      } else {
+        cartRepo.addItem(cartEntity, req.body.item)
+        cartEntity = cartRepo.updateCart(cartEntity)
+        baseResponse.response = await cartRepo.saveCart(cartEntity)
+        baseResponse.isSuccess = true
+        baseResponse.responseCode = 200
+      }
+    } else {
+      baseResponse.isSuccess = false
+      baseResponse.responseCode = 400
+      baseResponse.message = 'invalid cart id'
+    }
   } catch (e) {
     baseResponse.isSuccess = false
-    baseResponse.response = JSON.stringify(e)
+    baseResponse.response = e
   }
-  res.send(baseResponse)
+  res.send(baseResponse.responseCode, baseResponse)
 }
 
 /**
@@ -58,17 +69,26 @@ export let removeItem = async (req: Request, res: Response) => {
   let baseResponse: BaseResponse = new BaseResponse()
 
   try {
-    cartEntity = await cartRepo.getCartById(req.body.cart_id)
-    cartRepo.removeItem(cartEntity, req.body.item)
-    cartEntity = cartRepo.updateCart(cartEntity)
-    let result = await cartRepo.saveCart(cartEntity)
-    baseResponse.isSuccess = true
-    baseResponse.response = result
+    if (cartEntity != undefined) {
+      if (req.body.item == undefined) {
+        baseResponse.responseCode = 400
+        baseResponse.message = 'invalid item'
+      } else {
+        cartRepo.removeItem(cartEntity, req.body.item)
+        cartEntity = cartRepo.updateCart(cartEntity)
+        baseResponse.response = await cartRepo.saveCart(cartEntity)
+        baseResponse.responseCode = 200
+      }
+    } else {
+      baseResponse.isSuccess = false
+      baseResponse.responseCode = 400
+      baseResponse.message = 'invalid cart id'
+    }
   } catch (e) {
     baseResponse.isSuccess = false
-    baseResponse.response = JSON.stringify(e)
+    baseResponse.response = e
   }
-  res.send(baseResponse)
+  res.send(baseResponse.responseCode, baseResponse)
 }
 
 /**
@@ -87,11 +107,19 @@ export let fetchCart = async (req: Request, res: Response) => {
       result = await cartRepo.getAllCarts(req.query.page)
     }
 
-    baseResponse.isSuccess = true
-    baseResponse.response = result
+    if (result.length > 0) {
+      baseResponse.responseCode = 200
+      baseResponse.isSuccess = true
+      baseResponse.response = result
+    } else {
+      baseResponse.responseCode = 400
+      baseResponse.isSuccess = false
+      baseResponse.response = result
+      baseResponse.message = 'No cart found'
+    }
   } catch (e) {
     baseResponse.isSuccess = false
-    baseResponse.response = JSON.stringify(e)
+    baseResponse.response = e
   }
   res.send(baseResponse)
 }
@@ -111,7 +139,7 @@ export let deleteCart = async (req: Request, res: Response) => {
     baseResponse.response = result
   } catch (e) {
     baseResponse.isSuccess = false
-    baseResponse.response = JSON.stringify(e)
+    baseResponse.response = e
   }
   res.send(baseResponse)
 }
